create lunar location module
